Making common base class for Keras Neural Network tools.
